+++ risk-service/src/main/java/com/acquirerx/risk/risk/dto/RiskEventResponseDTO.java
@@ -2,6 +2,7 @@ package com.acquirerx.risk.risk.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -9,7 +10,7 @@ public class RiskEventResponseDTO {
 
     private Long riskEventId;
     private Long txnId;
-    private Double txnAmount;
+    private BigDecimal txnAmount;
     private Long ruleId;
     private String ruleName;
     private Integer score;
